backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
pydantic[email]>=2.9.0
alembic>=1.13.1
python-multipart>=0.0.9
